Implement SniperState and pass all tests #26
I had to change the way lost message is checked, because
of the different behavior when you joining and otherwise.

The book narative is too vague and it assumes certain structure, that
I don't think is presented in the previous chapters.
Nevermind, I was able to get things working and looking forward to clean
it.